requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docopt==0.6.2
fastavro==1.4.7
hdfs==2.7.0
idna==3.3
importlib-resources==5.4.0
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
zipp==3.6.0
mysql-connector-python==8.0.29
pkg-resources==0.0.0
protobuf==3.19.4
venv-pack==0.2.0

